Try importing from a .ts file, not a .tsx file
